refactor: add missingProvider helper
